Use the tableView to get the represented cell.
In the method before, the PXSimpleTableRow would end up retaining a cell that had scrolled out of the table, which would then get reused by another row. I doubt it's too safe to do this, and as we have all the info to get a cell from the table view, we can retrieve it like that.